requirements.txt
pytesseract==0.1.6
requests >= 2.0
future
line-bot-sdk
flask
pillow
gunicorn
pymongo
python-dateutil
python-telegram-bot
